requirements.txt
chromadb==1.0.21
faiss-cpu==1.8.0
langchain==0.3.27
langchain-community==0.3.30
langchain-core==0.3.80
langchain-ollama==0.3.8
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
llama-index-embeddings-openai==0.5.1
llama-index-llms-openai==0.5.6
openai==1.107.3
sentence-transformers==5.1.0
gradio<6
python-dotenv
python-docx
numpy<2
requests
pypandoc
pymupdf
pytest
torch
transformers
scikit-learn
